fix: reserve API misses for JSON responses (#158)
